Config: Add formal parameters
Previous design didn't allow to store parameters for a longer time -
they were parsed and consumed immediately. With this patch, a list
of parameters with their respective type must be given. This is
preparation to allow storing these parameters for a longer time, which
will allow setting properties at runtime basing on conditions given in
config file.

As a side effect, we can now generate command usage automatically :^)