Config+Player: Implement background transitions :^)
This uses a special AnimatableBackground class which is a pair of
texture and transition factor. It is compatible with AnimatableProperty.
After applying transition, it results in BlendedBackground class which
represents old and new background and is drawn on the screen.

The property callbacks tries to load background from file and stores it
in the map so that the file is not loaded in every transition frame. After
that, we can just call set_value_with_factor()! :^)