Event: Make SetTempo non-serializable for now
The serialize() function is commented out, so the SetTempo function is
not written correctly to file. This caused only delta time to be
written, resulting in an invalid MIDI file. This commit temporarily
disables serialization of SetTempo events until we figure out the
FIXME that blocks us from doing that.